feat : fsd/features/manual 1차 작업 완료
- 메뉴리스트 분리 작업 [이미지 아이템 분리]
- 에로우 화살표 [shard/component/arrow 분리]
- eslint default as 네이밍 룰 제거
