+++ client/.eslintrc.json
@@ -87,7 +87,8 @@
     "no-unused-expressions": [
       "error",
       { "allowShortCircuit": true, "allowTernary": true }
-    ]
+    ],
+    "no-restricted-exports": "off"
     // "indent": "off"
   }
 }
