+++ client/src/features/manual/components/Manual.tsx
