+++ client/src/features/manual/index.tsx
@@ -1,0 +1,1 @@
+export { default as Manual } from './components/Manual';
