+++ client/src/shared/components/arrow/ArrowDown.tsx
@@ -1,6 +1,7 @@
 import styled, { keyframes } from 'styled-components';
 
 import { BsArrowDown } from 'react-icons/bs';
+import { styleMixin } from '@/src/shared/styles';
 
 function ArrowDown() {
   return (
@@ -23,12 +24,10 @@ const bounce = keyframes`
 
 const ArrowContainer = styled.div`
   position: fixed;
-  bottom: 8rem;
+  ${styleMixin.Flex()};
+  margin-top: 5rem;
   left: 50%;
   transform: translateX(-50%);
-  display: flex;
-  justify-content: center;
-  align-items: center;
   animation: ${bounce} 2s infinite;
   opacity: 0.8;
   svg {
