+++ client/src/shared/components/arrow/index.ts
@@ -1,0 +1,1 @@
+export { default as ArrowDown } from './ArrowDown';
